+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- reflect-config.json generation capability
+- NativeHelperFacade support for typeReachable condition
+
 ## [1.2.0] - 2024-03-31
 
 ### Added